# tftypes: Reduce AttributePath and Value type compute and memory usage