--- a/tftypes/attribute_path.go
+++ b/tftypes/attribute_path.go
@@ -80,14 +80,20 @@ func (a *AttributePath) String() string {
 // AttributePaths are considered equal if they have the same number of steps,
 // the steps are all the same types, and the steps have all the same values.
 func (a *AttributePath) Equal(o *AttributePath) bool {
-	if len(a.Steps()) == 0 && len(o.Steps()) == 0 {
-		return true
+	if a == nil {
+		return o == nil || len(o.steps) == 0
 	}
-	if len(a.Steps()) != len(o.Steps()) {
+
+	if o == nil {
+		return len(a.steps) == 0
+	}
+
+	if len(a.steps) != len(o.steps) {
 		return false
 	}
-	for pos, aStep := range a.Steps() {
-		oStep := o.Steps()[pos]
+
+	for pos, aStep := range a.steps {
+		oStep := o.steps[pos]
 
 		if !aStep.Equal(oStep) {
 			return false
@@ -120,63 +126,112 @@ func (a *AttributePath) NewError(err error) error {
 	}
 }
 
-// LastStep returns the last step in the path. If the path was
-// empty, nil is returned.
+// LastStep returns the last step in the path. If the path is nil or empty, nil
+// is returned.
 func (a *AttributePath) LastStep() AttributePathStep {
-	steps := a.Steps()
-
-	if len(steps) == 0 {
+	if a == nil || len(a.steps) == 0 {
 		return nil
 	}
 
-	return steps[len(steps)-1]
+	return a.steps[len(a.steps)-1]
+}
+
+// NextStep returns the next step in the path. If the path is nil or empty, nil
+// is returned.
+func (a *AttributePath) NextStep() AttributePathStep {
+	if a == nil || len(a.steps) == 0 {
+		return nil
+	}
+
+	return a.steps[0]
 }
 
 // WithAttributeName adds an AttributeName step to `a`, using `name` as the
 // attribute's name. `a` is copied, not modified.
 func (a *AttributePath) WithAttributeName(name string) *AttributePath {
-	steps := a.Steps()
+	if a == nil {
+		return &AttributePath{
+			steps: []AttributePathStep{AttributeName(name)},
+		}
+	}
+
+	// Avoid re-allocating larger slice
+	steps := make([]AttributePathStep, len(a.steps)+1)
+	copy(steps, a.steps)
+	steps[len(steps)-1] = AttributeName(name)
+
 	return &AttributePath{
-		steps: append(steps, AttributeName(name)),
+		steps: steps,
 	}
 }
 
 // WithElementKeyString adds an ElementKeyString step to `a`, using `key` as
 // the element's key. `a` is copied, not modified.
 func (a *AttributePath) WithElementKeyString(key string) *AttributePath {
-	steps := a.Steps()
+	if a == nil {
+		return &AttributePath{
+			steps: []AttributePathStep{ElementKeyString(key)},
+		}
+	}
+
+	// Avoid re-allocating larger slice
+	steps := make([]AttributePathStep, len(a.steps)+1)
+	copy(steps, a.steps)
+	steps[len(steps)-1] = ElementKeyString(key)
+
 	return &AttributePath{
-		steps: append(steps, ElementKeyString(key)),
+		steps: steps,
 	}
 }
 
 // WithElementKeyInt adds an ElementKeyInt step to `a`, using `key` as the
 // element's key. `a` is copied, not modified.
 func (a *AttributePath) WithElementKeyInt(key int) *AttributePath {
-	steps := a.Steps()
+	if a == nil {
+		return &AttributePath{
+			steps: []AttributePathStep{ElementKeyInt(key)},
+		}
+	}
+
+	// Avoid re-allocating larger slice
+	steps := make([]AttributePathStep, len(a.steps)+1)
+	copy(steps, a.steps)
+	steps[len(steps)-1] = ElementKeyInt(key)
+
 	return &AttributePath{
-		steps: append(steps, ElementKeyInt(key)),
+		steps: steps,
 	}
 }
 
 // WithElementKeyValue adds an ElementKeyValue to `a`, using `key` as the
 // element's key. `a` is copied, not modified.
 func (a *AttributePath) WithElementKeyValue(key Value) *AttributePath {
-	steps := a.Steps()
+	if a == nil {
+		return &AttributePath{
+			steps: []AttributePathStep{ElementKeyValue(key)},
+		}
+	}
+
+	// Avoid re-allocating larger slice
+	steps := make([]AttributePathStep, len(a.steps)+1)
+	copy(steps, a.steps)
+	steps[len(steps)-1] = ElementKeyValue(key)
+
 	return &AttributePath{
-		steps: append(steps, ElementKeyValue(key.Copy())),
+		steps: steps,
 	}
 }
 
 // WithoutLastStep removes the last step, whatever kind of step it was, from
 // `a`. `a` is copied, not modified.
 func (a *AttributePath) WithoutLastStep() *AttributePath {
-	steps := a.Steps()
-	if len(steps) == 0 {
+	if a == nil || len(a.steps) == 0 {
 		return nil
 	}
+
 	return &AttributePath{
-		steps: steps[:len(steps)-1],
+		// Paths are immutable, so this should be safe without copying.
+		steps: a.steps[:len(a.steps)-1],
 	}
 }
 
@@ -296,7 +351,7 @@ type AttributePathStepper interface {
 // types need to use the AttributePathStepper interface to tell
 // WalkAttributePath how to traverse themselves.
 func WalkAttributePath(in interface{}, path *AttributePath) (interface{}, *AttributePath, error) {
-	if len(path.Steps()) < 1 {
+	if path == nil || len(path.steps) == 0 {
 		return in, path, nil
 	}
 	stepper, ok := in.(AttributePathStepper)
@@ -306,11 +361,11 @@ func WalkAttributePath(in interface{}, path *AttributePath) (interface{}, *Attri
 			return in, path, ErrNotAttributePathStepper
 		}
 	}
-	next, err := stepper.ApplyTerraform5AttributePathStep(path.Steps()[0])
+	next, err := stepper.ApplyTerraform5AttributePathStep(path.NextStep())
 	if err != nil {
 		return in, path, err
 	}
-	return WalkAttributePath(next, &AttributePath{steps: path.Steps()[1:]})
+	return WalkAttributePath(next, NewAttributePathWithSteps(path.steps[1:]))
 }
 
 func builtinAttributePathStepper(in interface{}) (AttributePathStepper, bool) {