--- a/tftypes/value_benchmark_test.go
+++ b/tftypes/value_benchmark_test.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package tftypes
 
 import "testing"